vdev: create ZAP for root vdev
And add it to the AVZ, this is not backwards compatible with older pools
due to an assertion in spa_sync() that verifies the number of ZAPs of
all vdevs matches the number of ZAPs in the AVZ.

Granted, the assertion described above only applies to #DEBUG builds.
Still, a new feature flag, com.klarasystems:avz_v2, has been introduced
to avoid the assertion altogether.